+++ tests/unit/validator/ValidatorTest.php
@@ -1,0 +1,330 @@
+<?php
+
+/**
+ * @copyright Frederic G. Østby
+ * @license   http://www.makoframework.com/license
+ */
+
+namespace mako\tests\unit\validator;
+
+use mako\i18n\I18n;
+use mako\tests\TestCase;
+use mako\validator\Validator;
+use Mockery;
+
+/**
+ * @group unit
+ */
+class ValidatorTest extends TestCase
+{
+	/**
+	 * Rule set spy
+	 *
+	 * @param  \mako\validator\Validator $validator Validator
+	 * @param  string                    $attribute Attribute name
+	 * @return array
+	 */
+	protected function attributeSpy(Validator $validator, string $attribute): array
+	{
+		return (function() use ($attribute)
+		{
+			return $this->$attribute;
+		})->bindTo($validator, Validator::class)();
+	}
+
+	/**
+	 *
+	 */
+	public function testExtend()
+	{
+		$validator = new Validator([], []);
+
+		$this->assertFalse(array_key_exists('foobar', $this->attributeSpy($validator, 'rules')));
+
+		$validator->extend('foobar', 'Foobar');
+
+		$this->assertTrue(array_key_exists('foobar', $this->attributeSpy($validator, 'rules')));
+
+		$this->assertSame('Foobar', $this->attributeSpy($validator, 'rules')['foobar']);
+	}
+
+	/**
+	 *
+	 */
+	public function testBasicRules()
+	{
+		$ruleSets =
+		[
+			'foo' => ['required', 'min_length(10)'],
+			'bar' => ['min_length(10)'],
+		];
+
+		$validator = new Validator([], $ruleSets);
+
+		$this->assertSame($ruleSets, $this->attributeSpy($validator, 'ruleSets'));
+	}
+
+	/**
+	 *
+	 */
+	public function testAddRules()
+	{
+		$validator = new Validator([], []);
+
+		$this->assertSame([], $this->attributeSpy($validator, 'ruleSets'));
+
+		$validator->addRules('foo', ['required', 'min_length(10)']);
+		$validator->addRules('foo', ['max_length(20)']);
+
+		$this->assertSame(['foo' => ['required', 'min_length(10)', 'max_length(20)']], $this->attributeSpy($validator, 'ruleSets'));
+	}
+
+	/**
+	 *
+	 */
+	public function testAddRulesIf()
+	{
+		$validator = new Validator([], []);
+
+		$this->assertSame([], $this->attributeSpy($validator, 'ruleSets'));
+
+		$validator->addRulesIf('foo', ['required', 'min_length(10)'], true);
+		$validator->addRulesIf('foo', ['max_length(20)'], function() { return true; });
+
+		$validator->addRulesIf('foo', ['required', 'min_length(10)'], false);
+		$validator->addRulesIf('foo', ['max_length(20)'], function() { return false; });
+
+		$this->assertSame(['foo' => ['required', 'min_length(10)', 'max_length(20)']], $this->attributeSpy($validator, 'ruleSets'));
+	}
+
+	/**
+	 *
+	 */
+	public function testWildcardEpansion()
+	{
+		$input =
+		[
+			'users' =>
+			[
+				['foo' => ''],
+				['foo' => ''],
+				['foo' => ''],
+			],
+		];
+
+		$ruleSets =
+		[
+			'users.*.foo' => ['required'],
+		];
+
+		$validator = new Validator($input, $ruleSets);
+
+		$expectedRuleSets =
+		[
+			'users.0.foo' => ['required'],
+			'users.1.foo' => ['required'],
+			'users.2.foo' => ['required'],
+		];
+
+		$this->assertSame($expectedRuleSets, $this->attributeSpy($validator, 'ruleSets'));
+
+		$expecptedOriginalFieldNames =
+		[
+			'users.0.foo' => 'users.*.foo',
+			'users.1.foo' => 'users.*.foo',
+			'users.2.foo' => 'users.*.foo',
+		];
+
+		$this->assertSame($expecptedOriginalFieldNames, $this->attributeSpy($validator, 'originalFieldNames'));
+	}
+
+	/**
+	 *
+	 */
+	public function testNestedWildcardEpansion()
+	{
+		$input =
+		[
+			'users' =>
+			[
+				['foo' => ['bar' => '']],
+				['foo' => ['bar' => '']],
+				['foo' => ['bar' => '']],
+			],
+		];
+
+		$ruleSets =
+		[
+			'users.*.*.*' => ['required'],
+		];
+
+		$validator = new Validator($input, $ruleSets);
+
+		$expectedRuleSets =
+		[
+			'users.0.foo.bar' => ['required'],
+			'users.1.foo.bar' => ['required'],
+			'users.2.foo.bar' => ['required'],
+		];
+
+		$this->assertSame($expectedRuleSets, $this->attributeSpy($validator, 'ruleSets'));
+
+		$expecptedOriginalFieldNames =
+		[
+			'users.0.foo.bar' => 'users.*.*.*',
+			'users.1.foo.bar' => 'users.*.*.*',
+			'users.2.foo.bar' => 'users.*.*.*',
+		];
+
+		$this->assertSame($expecptedOriginalFieldNames, $this->attributeSpy($validator, 'originalFieldNames'));
+	}
+
+	/**
+	 *
+	 */
+	public function testWildcardEpansionWithNoInput()
+	{
+		$ruleSets =
+		[
+			'users.*.foo' => ['required'],
+		];
+
+		$validator = new Validator([], $ruleSets);
+
+		$this->assertSame([], $this->attributeSpy($validator, 'ruleSets'));
+	}
+
+	/**
+	 *
+	 */
+	public function testValidateOnEmptyFieldWithRuleThatDoesntValidateEmptyInput()
+	{
+		$input =
+		[
+			'email' => '',
+		];
+
+		$ruleSets =
+		[
+			'email' => ['email'],
+		];
+
+		$validator = new Validator($input, $ruleSets);
+
+		$this->assertTrue($validator->isValid($errors1));
+
+		$this->assertFalse($validator->isInvalid($errors2));
+
+		$this->assertSame([], $validator->getErrors());
+
+		$this->assertSame($errors1, $validator->getErrors());
+
+		$this->assertSame($errors2, $validator->getErrors());
+	}
+
+	/**
+	 *
+	 */
+	public function testValidateOnEmptyFieldWithRuleThatValidatesEmptyInput()
+	{
+		$input =
+		[
+			'email' => '',
+		];
+
+		$ruleSets =
+		[
+			'email' => ['required'],
+		];
+
+		$expectedErrors = ['email' => 'The email field is required.'];
+
+		$validator = new Validator($input, $ruleSets);
+
+		$this->assertFalse($validator->isValid($errors1));
+
+		$this->assertTrue($validator->isInvalid($errors2));
+
+		$this->assertSame($expectedErrors, $validator->getErrors());
+
+		$this->assertSame($expectedErrors, $errors1);
+
+		$this->assertSame($expectedErrors, $errors2);
+	}
+
+	/**
+	 *
+	 */
+	public function testThatErrorMessagesForExpandedFieldsUseTheOriginalName()
+	{
+		$input =
+		[
+			'user' => ['email' => ''],
+		];
+
+		$ruleSets =
+		[
+			'user.*' => ['required'],
+		];
+
+		$validator = new Validator($input, $ruleSets);
+
+		$this->assertFalse($validator->isValid());
+
+		$this->assertTrue($validator->isInvalid());
+
+		$this->assertSame(['user.email' => 'The user.* field is required.'], $validator->getErrors());
+	}
+
+	/**
+	 *
+	 */
+	public function testThatParametersGetPassedToRule()
+	{
+		$input =
+		[
+			'foo' => '123456',
+		];
+
+		$ruleSets =
+		[
+			'foo' => ['max_length(4)'],
+		];
+
+		$validator = new Validator($input, $ruleSets);
+
+		$this->assertFalse($validator->isValid());
+
+		$this->assertTrue($validator->isInvalid());
+
+		$this->assertSame(['foo' => 'The value of the foo field must be at most 4 characters long.'], $validator->getErrors());
+	}
+
+	/**
+	 *
+	 */
+	public function testGetErrorMessageWithI18n()
+	{
+		$input =
+		[
+			'foo' => '123456',
+		];
+
+		$ruleSets =
+		[
+			'foo' => ['max_length(4)'],
+		];
+
+		$i18n = Mockery::mock(I18n::class);
+
+		$i18n->shouldReceive('has')->once()->andReturn(true);
+
+		$i18n->shouldReceive('get')->once()->andReturn('custom message');
+
+		$validator = new Validator($input, $ruleSets, $i18n);
+
+		$this->assertFalse($validator->isValid());
+
+		$this->assertSame(['foo' => 'custom message'], $validator->getErrors());
+	}
+}